Validation

+++ api/Controllers/CommentController.cs
@@ -21,22 +21,25 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var comments = await _commentRepo.GetAllAsync();
         var commentsDto = comments.Select(s => s.ToCommentDto());
         return Ok(commentsDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var comment = await _commentRepo.GetByIdAsync(id);
         if (comment == null) return NotFound();
         return Ok(comment.ToCommentDto());
     }
 
-    [HttpPost("{stockId}")]
+    [HttpPost("{stockId:int}")]
     public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
     {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var stock = await _stockRepo.StockExistsAsync(stockId);
         if (!stock) return BadRequest("Stock doesn't exist");
         var comment = commentDto.ToCommentFromCreateDto(stockId);
@@ -44,17 +47,19 @@ public class CommentController : ControllerBase
         return CreatedAtAction(nameof(GetById), new { id = comment.Id }, comment.ToCommentDto());
     }
 
-    [HttpPut("{id}")]
+    [HttpPut("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
     {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var comment = await _commentRepo.UpdateAsync(id, updateDto);
         if (comment == null) return NotFound();
         return Ok(comment.ToCommentDto());
     }
 
-    [HttpDelete("{id}")]
+    [HttpDelete("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var comment = await _commentRepo.DeleteAsync(id);
         if (comment == null) return NotFound();
         return NoContent();
